+++ lib/src/service/spaces/spaces_service.dart
@@ -22,6 +22,7 @@ import 'space_expansion.dart';
 import 'space_field.dart';
 import 'space_meta.dart';
 import 'space_state.dart';
+import 'topic_field.dart';
 
 abstract class SpacesService {
   /// Returns the new instance of [SpacesService].
@@ -66,6 +67,11 @@ abstract class SpacesService {
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
   ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
+  ///
   /// ## Endpoint Url
   ///
   /// - https://api.twitter.com/2/spaces/search
@@ -98,6 +104,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns a variety of information about a single Space specified by the
@@ -126,6 +133,11 @@ abstract class SpacesService {
   /// - [spaceFields]: This fields parameter enables you to select which
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
+  ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
   ///
   /// ## Endpoint Url
   ///
@@ -158,6 +170,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns details about multiple Spaces. Up to 100 comma-separated Spaces
@@ -186,6 +199,11 @@ abstract class SpacesService {
   /// - [spaceFields]: This fields parameter enables you to select which
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
+  ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
   ///
   /// ## Endpoint Url
   ///
@@ -218,6 +236,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns a list of user who purchased a ticket to the requested Space.
@@ -453,6 +472,11 @@ abstract class SpacesService {
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
   ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
+  ///
   /// ## Endpoint Url
   ///
   /// - https://api.twitter.com/2/spaces/by/creator_ids
@@ -487,6 +511,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 }
 
@@ -501,6 +526,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -512,6 +538,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -524,6 +551,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildResponse(
         await super.get(
@@ -533,6 +561,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -544,6 +573,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -554,6 +584,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -618,6 +649,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -628,6 +660,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,